+++ src/core/nodes/abstract-node.ts
@@ -51,6 +51,24 @@ export abstract class AbstractNode  {
     abstract setParent(newParent: ComplexNode, subpath: string): void
 }
 
+
+export function getRelativePathForNodes(base: ComplexNode, target: ComplexNode): string {
+    // PRE condition target is (a child of) base!
+    if (base.root !== target.root) fail(`Cannot calculate relative path: objects '${base}' and '${target}' are not part of the same object tree`)
+
+    const baseParts = splitJsonPath(base.path)
+    const targetParts = splitJsonPath(target.path)
+    let common = 0
+    for (; common < baseParts.length; common++) {
+        if (baseParts[common] !== targetParts[common])
+            break
+    }
+    // TODO: assert that no targetParts paths are "..", "." or ""!
+    return baseParts.slice(common).map(_ => "..").join("/")
+        + joinJsonPath(targetParts.slice(common))
+}
+
 import { IType } from "../../types/type"
-import { escapeJsonPath } from "../json-patch"
+import { escapeJsonPath, splitJsonPath, joinJsonPath } from "../json-patch"
 import { ComplexNode } from "./complex-node"
+import { fail } from "../../utils"
